Added Platform 3
Lilypad's serve as a way to jump onto the next platform. Another walkway added to allow player to jump back onto platform 2. Blockade added to platform 3 to prevent the players from progressing without finding weapons.

Platform 3 is a large area and will be the point of first combat after the the rocks are destroyed.